requirements_dev.txt
-e .
black==24.4.2
coverage==7.5.1
mypy==1.10.0
pdoc==14.4.0
pip==24.0
pre-commit==3.7.1
pytest-cov==5.0.0
pytest==8.2.0
ruff==0.4.4
pytest-aiohttp==0.24.0
